MODIFIED: Custom directories in 01Clase and 04Clase
